requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
pydantic
fastapi
uvicorn[standard]
pandas
numpy
python-dotenv
openai   # for OpenAI HTTP client

google-generativeai
requests
